requirements.txt
absl-py==0.10.0
attrs==20.2.0
git-python==1.0.3
gym==0.17.2
matplotlib==3.2.1
networkx==2.5
nltk==3.4.5
numpy==1.18.1
pandas==1.0.3
protobuf==3.11.3
pyparsing==2.4.7
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
revtok==0.0.3
scikit-image==0.14.2
scipy==1.4.1
sentencepiece==0.1.82
six==1.14.0
smmap==3.0.4
tensorboardX==2.0
